+++ test/test_api.rb
@@ -1,6 +1,7 @@
 require 'helper'
 
 class TestApi < Sidekiq::Test
+
   describe "stats" do
     before do
       Sidekiq.redis {|c| c.flushdb }
@@ -160,6 +161,8 @@ class TestApi < Sidekiq::Test
   end
 
   describe 'with an empty database' do
+    include Sidekiq::Util
+
     before do
       Sidekiq.redis {|c| c.flushdb }
     end
@@ -339,6 +342,23 @@ class TestApi < Sidekiq::Test
       assert_equal 0, r.size
     end
 
+    it 'can enumerate processes' do
+      odata = { 'pid' => 123, 'hostname' => hostname, 'key' => "#{hostname}:123", 'started_at' => Time.now.to_f - 15 }
+      time = Time.now.to_f
+      Sidekiq.redis do |conn|
+        conn.multi do
+          conn.sadd('processes', odata['key'])
+          conn.hmset(odata['key'], 'info', Sidekiq.dump_json(odata), 'busy', 10, 'beat', time)
+          conn.sadd('processes', 'fake:pid')
+        end
+      end
+
+      ps = Sidekiq::ProcessSet.new.to_a
+      assert_equal 1, ps.size
+      data = ps.first
+      assert_equal odata.merge('busy' => 10, 'beat' => time), data
+    end
+
     it 'can enumerate workers' do
       w = Sidekiq::Workers.new
       assert_equal 0, w.size
@@ -346,37 +366,36 @@ class TestApi < Sidekiq::Test
         assert false
       end
 
-      s = '12345'
-      data = Sidekiq.dump_json({ 'payload' => {}, 'queue' => 'default', 'run_at' => Time.now.to_i })
-      Sidekiq.redis do |c|
-        c.multi do
-          c.sadd('workers', s)
-          c.set("worker:#{s}", data)
-        end
+      key = "#{hostname}:#{$$}"
+      pdata = { 'pid' => $$, 'hostname' => hostname, 'started_at' => Time.now.to_i }
+      Sidekiq.redis do |conn|
+        conn.sadd('processes', key)
+        conn.hmset(key, 'info', Sidekiq.dump_json(pdata), 'busy', 0, 'beat', Time.now.to_f)
       end
 
-      assert_equal 1, w.size
-      w.each do |x, y|
-        assert_equal s, x
+      s = "#{key}:workers"
+      data = Sidekiq.dump_json({ 'payload' => {}, 'queue' => 'default', 'run_at' => Time.now.to_i })
+      Sidekiq.redis do |c|
+        c.hmset(s, '1234', data)
+      end
+
+      w.each do |p, x, y|
+        assert_equal key, p
+        assert_equal "1234", x
         assert_equal 'default', y['queue']
         assert_equal Time.now.year, Time.at(y['run_at']).year
       end
 
-      s = '12346'
+      s = "#{key}:workers"
       data = Sidekiq.dump_json({ 'payload' => {}, 'queue' => 'default', 'run_at' => (Time.now.to_i - 2*60*60) })
       Sidekiq.redis do |c|
         c.multi do
-          c.sadd('workers', s)
-          c.set("worker:#{s}", data)
-          c.set("worker:#{s}:started", Time.now.to_s)
-          c.sadd('workers', '123457')
+          c.hmset(s, '5678', data)
+          c.hmset("b#{s}", '5678', data)
         end
       end
 
-      assert_equal 3, w.size
-      count = w.prune
-      assert_equal 1, w.size
-      assert_equal 2, count
+      assert_equal ['1234', '5678'], w.map { |pid, tid, data| tid }
     end
 
     it 'can reschedule jobs' do